Move tableFilters and tableColumns in a hook and use it to get data.
